基础架构调整

+++ src/libs/libserver/util_time.h
@@ -1,28 +1,29 @@
 #pragma once
 
+#include <chrono>
 #include "common.h"
 #include "util_string.h"
 
-#include <chrono>
-
 namespace timeutil
 {
-    // uint64，单位：毫秒
-	typedef uint64 Time;  // milliseconds from 1971
-    // timeValue之后second秒之后的毫秒数，单位：毫秒
-	inline Time AddSeconds(Time timeValue, int second)
-	{
-		return timeValue + second * 1000;
-	}
-    // timeValue之后milliseconds毫秒之后的毫秒数，单位：毫秒
-	inline Time AddMilliseconds(Time timeValue, int milliseconds)
-	{
-		return timeValue + milliseconds;
-	}
-    // 获取当前时间，单位：毫秒
-	inline std::string NowToString()
-	{
-		auto timeValue = std::chrono::time_point_cast<std::chrono::milliseconds>(std::chrono::system_clock::now());
+    typedef uint64 Time; // 毫秒
+    
+    // 添加n秒之后的毫秒数
+    inline Time AddSeconds(Time timeValue, int second)
+    {
+        return timeValue + (uint64)second * 1000;
+    }
+
+    // 添加n毫秒之后的毫秒数
+    inline Time AddMilliseconds(Time timeValue, int milliseconds)
+    {
+        return timeValue + milliseconds;
+    }
+
+    // 获取当前时间的字符串
+    inline std::string NowToString()
+    {
+        auto timeValue = std::chrono::time_point_cast<std::chrono::milliseconds>(std::chrono::system_clock::now());
 #if ENGINE_PLATFORM != PLATFORM_WIN32
 		auto tt = std::chrono::system_clock::to_time_t(timeValue);
 		struct tm* ptm = localtime(&tt);
@@ -34,5 +35,5 @@ namespace timeutil
 		std::string rs = strutil::format("%d-%02d-%02d %02d:%02d:%02d", tm.tm_year + 1900, tm.tm_mon + 1, tm.tm_mday, tm.tm_hour, tm.tm_min, tm.tm_sec);
 #endif
 		return rs;
-	}
+    }
 }
